refactor: simplify invoice example module names
