+++ examples/invoice-triage/bin/watch.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env -S node --import tsx
-import { watchDirectory } from "../src/invoice-watcher"
-import { option, runtimeExtensions, runtimeTags, splitRuntimeArgs } from "./invoice-runtime"
+import { watchDirectory } from "../src/watcher"
+import { option, runtimeExtensions, runtimeTags, splitRuntimeArgs } from "./runtime"
 
 const { runtime, rest } = splitRuntimeArgs(process.argv.slice(2))
