+++ examples/invoice-triage/src/database.ts
@@ -2,7 +2,7 @@ import { asc, eq, inArray, sql } from "drizzle-orm"
 import { drizzle, type NodePgDatabase } from "drizzle-orm/node-postgres"
 import { Pool, type PoolConfig } from "pg"
 import { tag } from "@pumped-fn/lite"
-import { operationalFault } from "./invoice-errors"
+import { operationalFault } from "./errors"
 import {
   completedReport,
   databaseMigrations,
@@ -11,11 +11,11 @@ import {
   type DatabaseMigrationRecord,
   type DatabaseMigrationReport,
   type DatabaseMigrationStatus,
-} from "./invoice-migrations"
-import * as schema from "./invoice-schema"
-import { auditEvents, pendingInvoices, schemaMigrations, storedInvoices } from "./invoice-schema"
-import type { AuditAction, AuditEvent } from "./invoice-audit"
-import type { Invoice, SaveInvoiceInput, StoredInvoice } from "./invoice-types"
+} from "./migrations"
+import * as schema from "./schema"
+import { auditEvents, pendingInvoices, schemaMigrations, storedInvoices } from "./schema"
+import type { AuditAction, AuditEvent } from "./audit"
+import type { Invoice, SaveInvoiceInput, StoredInvoice } from "./types"
 
 export interface SaveInvoiceRecord extends SaveInvoiceInput {
   importedAt: string
