+++ examples/invoice-triage/src/intake.ts
@@ -1,9 +1,9 @@
 import { controller, flow, ParseError, tags } from "@pumped-fn/lite"
 import { logging } from "@pumped-fn/lite-extension-logging"
 import { step } from "@pumped-fn/sdk"
-import type { AuditEvent } from "./invoice-audit"
-import { clock, database, intakeLines } from "./invoice-runtime"
-import { importBatch } from "./invoice-triage"
+import type { AuditEvent } from "./audit"
+import { clock, database, intakeLines } from "./runtime"
+import { importBatch } from "./triage"
 import {
   enqueueInput,
   type EnqueueInput,
@@ -11,7 +11,7 @@ import {
   type IntakeSummary,
   type Invoice,
   type StoredInvoice,
-} from "./invoice-types"
+} from "./types"
 
 export const enqueue = flow({
   name: "invoice.enqueue",
