+++ examples/invoice-triage/src/server.ts
@@ -1,24 +1,24 @@
 import { createScope, type Lite } from "@pumped-fn/lite"
 import { serve } from "@hono/node-server"
 import { Hono, type Context } from "hono"
-import { prepareDatabase } from "./invoice-database-operations"
+import { prepareDatabase } from "./database-lifecycle"
 import {
   enqueue,
   ingest,
   listAuditEvents,
   listPendingInvoices,
   listStoredInvoices,
-} from "./invoice-intake"
+} from "./intake"
 import {
   dailyReport,
   dailyReportJob,
   sendReminders,
   sendRemindersJob,
   watchReviewQueue,
-} from "./invoice-reporting"
-import { databaseStartup } from "./invoice-runtime"
-import { importBatch } from "./invoice-triage"
-import type { DatabaseStartupMode } from "./invoice-migrations"
+} from "./reporting"
+import { databaseStartup } from "./runtime"
+import { importBatch } from "./triage"
+import type { DatabaseStartupMode } from "./migrations"
 
 const invalidJson = Symbol("invalid-json")
